+++ lib/strava/api/endpoints/activities.rb
@@ -36,6 +36,21 @@ module Strava
         end
 
         #
+        # List activity photos.
+        #
+        # @option options [String] :id
+        #   Activity id.
+        # @option options [Integer] :page
+        #   Page number.
+        # @option options [Integer] :per_page
+        #   Number of items per page. Defaults to 30.
+        #
+        def activity_photos(id_or_options, options = {}, &block)
+          id, options = parse_args(id_or_options, options)
+          paginate "activities/#{id}/photos", options, Strava::Models::Photo, &block
+        end
+
+        #
         # List activity kudoers.
         #
         # @option options [String] :id